Use sdk to handle console networking
