+++ app/domain/SlpNetworking.js
@@ -1,0 +1,166 @@
+/*
+
+Copyright (C) 2007 Free Software Foundation, Inc. <https://fsf.org/>
+
+This program is free software; you can redistribute it and/or
+modify it under the terms of the GNU Lesser General Public
+License as published by the Free Software Foundation; either
+version 3 of the License, or (at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public License
+along with this program; if not, write to the Free Software Foundation,
+Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
+
+*/
+
+
+import _ from 'lodash';
+
+import { ConsoleConnection, ConnectionStatus } from '@slippi/sdk';
+
+import { store } from '../index';
+import { connectionStateChanged } from '../actions/console';
+import DolphinManager from './DolphinManager';
+import SlpFileWriter from './SlpFileWriter';
+
+export default class SlpNetworking {
+  static connectionCount = 0;
+
+  constructor(settings = {}) {
+    SlpNetworking.connectionCount += 1;
+
+    this.id = SlpNetworking.connectionCount;
+    this.ipAddress = settings.ipAddress;
+    this.port = settings.port;
+    this.targetFolder = settings.targetFolder;
+    this.obsIP = settings.obsIP;
+    this.obsSourceName = settings.obsSourceName;
+    this.obsPassword = settings.obsPassword;
+    this.isRealTimeMode = settings.isRealTimeMode;
+    this.isRelaying = settings.isRelaying;
+
+    this.isMirroring = false;
+    this.connectionsByPort = [];
+    this.clientsByPort = [];
+    this.connectionStatus = ConnectionStatus.DISCONNECTED;
+    this.connDetails = this.getDefaultConnDetails();
+    this.consoleConn = null;
+
+    // A connection can mirror its received gameplay
+    this.dolphinManager = new DolphinManager(`mirror-${this.id}`, { mode: 'mirror' });
+
+    // Initialize SlpFileWriter for writting files
+    const slpSettings = {targetFolder: this.targetFolder, 
+      onFileStateChange: this.fileStateChangeHandler, 
+      obsIP: this.obsIP, obsSourceName: this.obsSourceName,
+      obsPassword: this.obsPassword, id: this.id,
+      isRelaying: this.isRelaying,
+    }
+    this.slpFileWriter = new SlpFileWriter(slpSettings);
+  }
+
+  forceConsoleUiUpdate() {
+    store.dispatch(connectionStateChanged());
+  }
+  
+  fileStateChangeHandler = () => {
+    this.forceConsoleUiUpdate();
+  }
+
+  getDefaultConnDetails() {
+    return {
+      gameDataCursor: Uint8Array.from([0, 0, 0, 0, 0, 0, 0, 0]), 
+      consoleNick: "unknown", 
+      version: "",
+      clientToken: 0,
+    };
+  }
+
+  getSettings() {
+    return {
+      id: this.id,
+      ipAddress: this.ipAddress,
+      port: this.port,
+      targetFolder: this.targetFolder,
+      obsIP: this.obsIP,
+      obsSourceName: this.obsSourceName,
+      obsPassword: this.obsPassword,
+      isRealTimeMode: this.isRealTimeMode,
+      isRelaying: this.isRelaying,
+      consoleNick: this.connDetails.consoleNick,
+    };
+  }
+
+  editSettings(newSettings) {
+    // If data is not provided, keep old values
+    this.ipAddress = newSettings.ipAddress || this.ipAddress;
+    this.port = newSettings.port || this.port;
+    this.targetFolder = newSettings.targetFolder || this.targetFolder;
+    this.obsIP = newSettings.obsIP || this.obsIP;
+    this.obsSourceName = newSettings.obsSourceName || this.obsSourceName;
+    this.obsPassword = newSettings.obsPassword || this.obsPassword;
+    this.isRealTimeMode = _.defaultTo(newSettings.isRealTimeMode, this.isRealTimeMode);
+    this.isRelaying = _.defaultTo(newSettings.isRelaying, this.isRelaying);
+  }
+
+  getDolphinManager() {
+    return this.dolphinManager;
+  }
+
+  connect() {
+    // Update dolphin manager settings
+    const connectionSettings = this.getSettings();
+    this.slpFileWriter.updateSettings(connectionSettings);
+    this.slpFileWriter.connectOBS();
+    this.dolphinManager.updateSettings(connectionSettings);
+
+    console.log("Connecting");
+    if (!this.consoleConn) this.consoleConn = new ConsoleConnection();
+    this.consoleConn.connect(this.ipAddress, this.port);
+    this.consoleConn.on("handshake", (details) => {
+      console.log(details);
+      this.connDetails = {...this.connDetails, ...details};
+      console.log(this.connDetails);
+    });
+    this.consoleConn.on("statusChange", (status) => {
+      this.connectionStatus = status;
+      this.forceConsoleUiUpdate();
+    });
+    this.consoleConn.on("data", (data) => this.handleReplayData(data));
+  }
+
+  disconnect() {
+    console.log("Disconnect request");
+
+    this.slpFileWriter.disconnectOBS();
+    this.consoleConn.disconnect();
+
+    this.forceConsoleUiUpdate();
+  }
+
+  handleReplayData(data) {
+    const result = this.slpFileWriter.handleData(data);
+    if (result.isNewGame) {
+      const curFilePath = this.slpFileWriter.getCurrentFilePath();
+      this.dolphinManager.playFile(curFilePath, false);
+    }
+  }
+
+  async startMirroring() {
+    try {
+      console.log("Mirroring start");
+      this.isMirroring = true;
+      this.forceConsoleUiUpdate();
+      await this.dolphinManager.startPlayback();
+    } finally {
+      console.log("Mirroring end");
+      this.isMirroring = false;
+      this.forceConsoleUiUpdate();
+    }
+  }
+}
